Authenticate to Tableau Server or Tableau Cloud

This function authenticates to a Tableau Server or Tableau Cloud using a personal access token (PAT).

Usage

authenticate_PAT(
  pat_name = tableau_pat_name(),
  pat_secret = tableau_pat_secret(),
  content_url = tableau_content_url(),
  base_url = tableau_base_url(),
  api_version = 3.4
)

Arguments

pat_name

The name of the personal access token (PAT). Defaults to the `TABLEAU_PAT_NAME` environment variable.

pat_secret

The secret of the personal access token (PAT). Defaults to the `TABLEAU_PAT_SECRET` environment variable.

content_url

The URL of the content to authenticate. Defaults to the `TABLEAU_CONTENT_URL` environment variable. For Tableau Server, this is typically the URL of the Tableau server followed by the site ID. For Tableau Cloud, this is usually the URL of the Tableau cloud workbook.

base_url

The base URL of the Tableau server or Tableau cloud. Defaults to the `TABLEAU_BASE_URL` environment variable. For Tableau Server, this is usually the URL of the Tableau server. For Tableau Cloud, this is usually the URL of the Tableau cloud, and it must contain the pod name, such as 10az, 10ay, or us-east-1. For example, the base URL to sign in to a site in the 10ay pod would be: https://10ay.online.tableau.com.

api_version

The API version to use. Default is 3.4.

Value

A list containing the base URL, the access token, the site ID, and the user ID.

See also